[Refactor] alert 관련 메서드 extension으로 분리
